izmene

+++ ingzanja/src/main/java/rs/ftn/ingzanja/dto/CaseBasedInfo.java
@@ -9,7 +9,7 @@ public class CaseBasedInfo {
     private List<String> simptomi;
 
 
-    public int getiDPregleda() {
+    public Long getiDPregleda() {
         return iDPregleda;
     }
